requirements.txt
google-cloud-aiplatform
google-genai
google-auth
gspread
pandas
pytest
